Fix the formalisation of `imo_1961_p1`
The previous version was incorrect because the `Nat` typing of the variables in the sets was incorrect, these should be reals.
Additionally, the question also asks to solve the system of equations for `x`, `y`, `z`.
Solution for this is take from [AoPS Solution 1](https://artofproblemsolving.com/wiki/index.php/1961_IMO_Problems/Problem_1).

PiperOrigin-RevId: 720183822